Handle end of history correctly for udbserver backend.
This changeset fixes a bug in which, after continuing to the end of a loaded
LiveRecorder recording, it became impossible to reverse into the debug session.

By using the same technique already implemented for rr support it is now
possible to reach the end of time and then continue debugging. In addition, the
exit code (where available) be reported correctly - although it will be reported
as zero at the end of time if no exit was recorded.

A workaround for apparent SIGCHILD handling (actually SIGSTOP, due to how the
gdbserial protocol encodes signals) is removed as it is not required with this
new implementation.