Use the executable from the recording as the debuggee path.
This updates the start-up of the replay session to obtain the local path to the
debuggee binary, which Delve wants to look up symbol information.

By doing this, it's possible to debug a recording even when the original binary
is no longer present or was at a different path on the machine that created the
recording.